finished R script

+++ R/scripts/create_matches_table.R
@@ -21,7 +21,26 @@ if (!dir.exists(date_file)){
 
 
 tournement_links<-  c("https://www.rugbypass.com/autumn-nations-cup/matches/",
-                      "https://www.rugbypass.com/premiership/matches/")
+                      "https://www.rugbypass.com/premiership/matches/",
+                      "https://www.rugbypass.com/six-nations/matches/",
+                      "https://www.rugbypass.com/the-rugby-championship/matches/",
+                      "https://www.rugbypass.com/internationals/matches/2019/",
+                      "https://www.rugbypass.com/internationals/matches/2020/",
+                      "https://www.rugbypass.com/premiership-cup/matches/",
+                      "https://www.rugbypass.com/european-champions-cup/matches/",
+                      "https://www.rugbypass.com/challenge-cup/matches/2019-2020/",
+                     "https://www.rugbypass.com/pro-14/matches/",
+                     "https://www.rugbypass.com/pro-14-rainbow-cup/matches/",
+                     "https://www.rugbypass.com/rainbow-cup-south-africa/matches/",
+                    "https://www.rugbypass.com/super-rugby-trans-tasman/matches/",
+                    "https://www.rugbypass.com/super-rugby-australia/matches/",
+                    "https://www.rugbypass.com/super-rugby-unlocked/matches/",
+                    "https://www.rugbypass.com/super-rugby/matches/",
+                    "https://www.rugbypass.com/mitre-10-cup/matches/",
+                    "https://www.rugbypass.com/currie-cup/matches/",
+                    "https://www.rugbypass.com/rugby-world-cup/matches/",
+                    "https://www.rugbypass.com/top-14/matches/"
+                    )
 
 
 for (tournement_link in tournement_links){
@@ -41,10 +60,16 @@ for (tournement_link in tournement_links){
   
   csv_file <- file.path(date_file, paste(comp_name, ".csv", sep = ""))
   
-
+if(file.exists( csv_file )){
+  read_csv(csv_file,
+           col_types = cols(season = "c")) %>% 
+    bind_rows(table) %>% 
+    distinct(rugby_pass_url, .keep_all = TRUE) %>% 
+    write_csv(csv_file)
+} else {
  table %>% 
       write_csv(csv_file)
-
+}
 
 
 }
